X2-4697 Extract UpcomingDatePicker into a new component
