+++ src/components/DatePicker/DatePicker.jsx
@@ -11,6 +11,7 @@ import { Day } from "./Day";
 import { MonthYearSelector } from "./MonthYearSelector";
 import { NavbarElement } from "./NavbarElement";
 import { RelativeDateRange } from "./RelativeDateRange";
+import { UpcomingDatePicker } from "./UpcomingDatePicker";
 
 const variants = {
     single: "single",
@@ -144,38 +145,12 @@ export const DatePicker = ({
         <>
             <div className="flex">
                 {upcomingDates ? (
-                    <div className="rounded-l-lg border-r border-gray pt-8">
-                        <p className="mb-2 px-6 text-lg font-bold">Upcoming</p>
-                        {upcomingDates?.length > 0 ? (
-                            <div className="mt-5">
-                                {upcomingDates?.map((date) => {
-                                    const isSameDay = dayjs(date).isSame(dayjs(value), "day");
-                                    const key = dayjs(date).format();
-                                    return (
-                                        <div
-                                            key={key}
-                                            value
-                                            className={clsx(
-                                                "mx-6 mt-3 flex min-w-40 cursor-pointer items-center justify-center",
-                                                "rounded border border-gray py-3 hover:border-blue hover:bg-blue hover:text-white",
-                                                { "border-blue bg-blue text-white": isSameDay },
-                                            )}
-                                            onClick={(event) => {
-                                                handleDayClick(date, {}, event);
-                                                handleMonthChange(date);
-                                            }}
-                                        >
-                                            {dayjs(date).format("ddd DD MMMM")}
-                                        </div>
-                                    );
-                                })}
-                            </div>
-                        ) : (
-                            <div className="mx-6 mt-7 max-w-40 items-center justify-center rounded bg-yellow-lighter p-3">
-                                There is no future availability for this product.
-                            </div>
-                        )}
-                    </div>
+                    <UpcomingDatePicker
+                        upcomingDates={upcomingDates}
+                        value={value}
+                        onDayClick={handleDayClick}
+                        onMonthChange={handleMonthChange}
+                    />
                 ) : null}
 
                 <DayPicker
