+++ src/stories/DataDisplay/DateRangePicker.stories.js
@@ -47,7 +47,7 @@ export const RelativeDateRanges = () => {
     const [value, setValue] = useState({ from: new Date("2022-03-03"), to: new Date("2022-04-08") });
 
     return (
-        <div className="flex w-3/4 flex-col">
+        <div className="flex w-3/4 flex-col border border-gray-dark">
             <DatePicker
                 shouldShowYearPicker
                 shouldShowRelativeRanges
